% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efr_aggregation.R
\name{applyBasinData}
\alias{applyBasinData}
\title{Apply data for each basin to all cells of that basin}
\usage{
applyBasinData(data, endcell)
}
\arguments{
\item{data}{array with information for each basin}

\item{endcell}{lpjml array giving the index of the final drainage cell
for each cell}
}
\value{
array with data for each grid cell, based on the final drainage cell
}
\description{
Apply data for each basin to all cells of that basin
}
