% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_nyear_window.R
\name{average_nyear_window}
\alias{average_nyear_window}
\title{Calculate averages (mean) for defined window sizes}
\usage{
average_nyear_window(
  x,
  nyear = NULL,
  moving_average = FALSE,
  interpolate = FALSE
)
}
\arguments{
\item{x}{LPJmL output array with \code{dim(x)=c(ncell, months, years)}}

\item{nyear}{integer, if supplied it defines the years for each window to
be averaged over in \code{dim(x)[3]}. If \code{nyear == 1} values are used directly
(instead of calculating an average). nyear has to be smaller than
\code{dim(x)[3]} and \code{dim(x)[3]} has to be a multiply of nyear.
Defaults to \code{NULL}}

\item{moving_average}{logical. If \code{TRUE} moving average is computed. start
and end are interpolated using spline interpolation.}

\item{interpolate}{logical. If \code{TRUE} and nyear is defined (with
\code{moving_average == FALSE} years are interpolated (spline) to return array
with same dimensions as \code{x} (mainly\code{dim(x)[3]} -> years).}
}
\value{
array with same amount of cells and months as x. 3rd dimension is
defined by nyear, basically \code{dim(x)[3]/nyear} or equal to dim(x)\link{3} if
\code{moving_average == TRUE} or \code{interpolate == TRUE}
}
\description{
Define window sizes (nyear) to be used to calculate averages (mean) for each
window (\code{dim(x)[3] / nyear}). Instead of discrete windows, also moving
averages can be computed as well as years inbetween interpolated.
}
