% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosphere_status.R
\name{biosphere_status}
\alias{biosphere_status}
\title{Status calculation of the biosphere integrity boundary.}
\usage{
biosphere_status(
  files_scenario,
  files_reference,
  spatial_scale = "subglobal",
  time_span_scenario = as.character(1982:2011),
  time_span_reference = NULL,
  approach = "stenzel2023",
  time_series_avg = NULL,
  config_args = list(),
  thresholds = NULL,
  path_baseline,
  time_span_baseline = time_span_scenario,
  npp_threshold = 20,
  biocol_option = "only_above_zero",
  eurasia = TRUE,
  ...
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. All needed files need to be
provided. E.g.: list(grid = "/temp/grid.bin.json",
npp = "/temp/npp.bin.json"). Handled via \code{\link{calc_status}}.}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the reference NPP, HANPP should be compared against.
In this case only NPP is required. list(npp = "/temp/npp.bin.json").}

\item{spatial_scale}{character string indicating spatial resolution
either "grid", "subglobal" or "global"}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as character string}

\item{time_span_reference}{time span to be used for the reference run,
defined as character string, e.g. \code{as.character(1901:1930)}.}

\item{approach}{approach (character string) to be used , currently available
approach is \code{"stenzel2023"}}

\item{time_series_avg}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for one status
calculation, for \code{1} the whole time span is used to calculate a status time
series.}

\item{config_args}{list of arguments to be passed on from the model
configuration.}

\item{thresholds}{named character string with thresholds to be used to
define the lower end of safe, increasing risk and high risk zone,
e.g. c(holocene = 0.0, pb = 0.1, highrisk = 0.2). If set to NULL, default
values from metric_files.yml will be used.}

\item{path_baseline}{character string with path to outputs for the baseline
run, file names are taken from files scenario.}

\item{time_span_baseline}{time span to be used for the baseline run, defined
as a character vector, e.g. \code{as.character(1901:1930)}. Can differ in offset
and length from \code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario}
is used}

\item{npp_threshold}{lower threshold for npp (to mask out non-lu areas
according to Haberl et al. 2007). Below BioCol will be set to 0.
(default: 20 gC/m2)}

\item{biocol_option}{which biocol values to use for aggregation. options:
netsum, only_above_zero, abs}

\item{eurasia}{logical. If \code{spatial_scale} = \code{"subglobal"} merge continents
Europe and Asia to avoid arbitrary biome cut at europe/asia border.
Defaults to \code{TRUE}}

\item{...}{arguments forwarded to \code{\link{classify_biomes}}}
}
\value{
Object of class \code{control_variable} with the boundary status of the
biosphere integrity boundary.
}
\description{
Biosphere status calculation based on BioCol (HANPP) from a baseline run
(with potential natural vegetation) and a scenario run (actual land use)
of LPJmL, both within the \code{time_span_scenario}.
Additionally a separate reference NPP file (e.g. from a Holocene run) can be
supplied with \code{files_reference} = list(npp = "path/to/npp.bin.json"),
which will use time_span_reference, or file index years 3:32 if
time_span_reference is not supplied.
}
\examples{
\dontrun{
boundary_status <- calc_status(
  boundary = "biosphere",
  config_scenario = "path/to/config_scenario.json",
  config_reference = "path/to/config_reference.json",
  spatial_scale = "global",
  time_span_scenario = 1901:2019,
  time_span_reference = 1901:1930,
  approach = "stenzel2023",
  path_baseline = "path/to/baseline_outputs"
)
}
}
