% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries-package.R
\docType{package}
\name{boundaries-package}
\alias{boundaries}
\alias{boundaries-package}
\title{boundaries: Functions to calculate the Planetary boundary status based on LPJmL simulation data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide calculation & visualizations functions for planetary boundaries.
}
\author{
\strong{Maintainer}: Jannes Breier \email{jannesbr@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9055-6904}{ORCID})

Authors:
\itemize{
  \item Fabian Stenzel \email{stenzel@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5109-0048}{ORCID})
  \item Johanna Braun \email{braun@pik-potsdam.de} (\href{https://orcid.org/0000-0002-0000-0000}{ORCID})
}

}
\keyword{internal}
