% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bluewater.R
\name{calc_bluewater}
\alias{calc_bluewater}
\title{Calculate bluewater planetary boundary status}
\usage{
calc_bluewater(
  path,
  method = "gerten2020",
  time_span = c(1982, 2011),
  bin_size = NULL,
  start_year = 1901,
  end_year = 2011
)
}
\arguments{
\item{path}{}

\item{method}{}

\item{time_span}{}

\item{bin_size}{}
}
\description{
Calculate ...
}
\examples{
\dontrun{
 calc_bluewater(path)
}

}
