% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_efrs.R
\name{calc_efrs}
\alias{calc_efrs}
\title{Calculate environmental flow requirements (EFRs)}
\usage{
calc_efrs(x, approach = "vmf")
}
\arguments{
\item{x}{discharge array with \code{dim(x)=c(cell, month, year)}. If no year
dimension is present, the function assumes a single year.}

\item{approach}{EFR approach to be used , available methods are \code{c("vmf", "q10q50")} based on a modified version of
\href{https://doi.org/10.5194/hess-18-5041-2014}{Pastor et al. 2014}
and c(vmf_min", "vmf_max") as suggested by
\href{https://doi.org/10.1038/s41893-019-0465-1}{Gerten et al. 2020}
and \href{https://doi.org/10.1126/science.1259855}{Steffen et al. 2015}}
}
\value{
EFRs with same unit as \code{x} (discharge). Cell and month dimensions are
preserved, and the year dimension is set to 1 (as EFRs are calculated
based on the whole provided period of years).
}
\description{
Calculate environmental flow requirements (EFRs) based on the number of years
of \code{dim(x)[3]}.
}
\examples{
\dontrun{
# basic example
efrs <- calc_efrs(x = discharge, approach = "vmf")

dim(efrs)
# c(67420, 12, 1)

}
}
