% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_greenwater_status.R
\name{calc_greenwater_status}
\alias{calc_greenwater_status}
\title{Calculate the planetary boundary status for the greenwater boundary}
\usage{
calc_greenwater_status(
  files_scenario,
  files_reference,
  time_span_scenario = as.character(1982:2011),
  time_span_reference = NULL,
  method = "wang-erlandsson2022",
  avg_nyear_args = list()
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. All needed files are
provided in XXX. E.g.: list(leaching = "/temp/leaching.bin.json")}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the reference LPJmL run. All needed files are
provided in XXX. E.g.: list(leaching = "/temp/leaching.bin.json"). If not
needed for the applied method, set to NULL.}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{method}{method (character string) to be used , currently available
method is \code{c("wang-erlandsson2022")} based on
\href{https://doi.org/10.1038/s43017-022-00287-8}{Wang-Erlandsson et al. 2022}.}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info). To be used for
time series analysis}
}
\description{
Calculate the PB status for the greenwater (former freshwater) boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
 calc_greenwater_status(path_scenario, path_reference)
}

}
