% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lsc_status.R
\name{calc_lsc_status}
\alias{calc_lsc_status}
\title{Calculate the planetary boundary status for the land-system change boundary}
\usage{
calc_lsc_status(
  files_scenario,
  files_reference,
  time_span_scenario = as.character(1982:2011),
  time_span_reference = NULL,
  lsc_spatial_scale = "biome",
  eurasia = TRUE,
  lsc_thresholds = list(temperate = list(lower = 0.5, upper = 0.7), boreal = list(lower =
    0.15, upper = 0.4), tropical = list(lower = 0.15, upper = 0.4)),
  avg_nyear_args = list(),
  ...
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. All needed files are
provided in XXX. E.g.: list(leaching = "/temp/leaching.bin.json")}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the reference LPJmL run. All needed files are
provided in XXX. E.g.: list(leaching = "/temp/leaching.bin.json"). If not
needed for the applied method, set to NULL.}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as a character string, e.g. \code{as.character(1982:2011)} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{as.character(1901:1930)}. Can differ in offset
and length from \code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario}
is used}

\item{lsc_spatial_scale}{character. Spatial resolution, available options
are \code{"biome"} (default) and \code{"grid"}}

\item{eurasia}{logical. If \code{lsc_spatial_scale} = \code{"biome"} merge continents
Europe and Asia to avoid arbitrary biome cut at europe/asia border.
Defaults to \code{TRUE}}

\item{lsc_thresholds}{list with deforestation thresholds for defining safe,
increasing risk and high risk zone for temperate/temperate/boreal forest
biomes, Default based on Steffen et al. 2015
(https://doi.org/10.1126/science.1259855):
list(boreal = list(lower = 0.5, upper = 0.7),
temperate = list(lower = 0.15, upper = 0.4),
tropical = list(lower = 0.15, upper = 0.4))
lower = threshold between safe zone and increasing risk zone (e.g. 50\% for
boreal forest with default value)
upper = threshold between increasing risk and high risk zone}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info).
To be used for time series analysis}

\item{...}{arguments forwarded to \link\href{classify_biomes}{boundaries}}
}
\description{
Calculate the PB status for the LSC (land-system change) boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
 calc_lsc_status(files_scenario, files_reference)
}

}
