% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nitrogen_status.R
\name{calc_nitrogen_status}
\alias{calc_nitrogen_status}
\title{Calculate the planetary boundary status for the nitrogen boundary}
\usage{
calc_nitrogen_status(
  path_scenario,
  path_reference,
  time_span_scenario = c(1982, 2011),
  time_span_reference = NULL,
  method = "braun2022",
  temporal_resolution = "annual",
  cut_arid = 0.2,
  cut_runoff = 0,
  with_groundwater_denit = TRUE,
  prefix_monthly_output = "",
  avg_nyear_args = list(),
  start_year = 1901
)
}
\arguments{
\item{path_scenario}{output directory (character string) of the scenario
LPJmL run where binary files (soon with metafiles) are written}

\item{path_reference}{output directory (character string) of the reference
\itemize{
\item if used - \verb{method == "braun2022_minusref}
}}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used if scenario run used, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{method}{method (character string) to be used , currently available
method is \code{"braun2022"} based on unpublished suggestion by Johanna Braun.
Second method option is \code{"braun2022_minusref"} to subtract reference run
output}

\item{temporal_resolution}{character. Temporal resolution, available options
are \code{"annual"} (default) and \code{"monthly"}}

\item{cut_arid}{double. Exclude boundary calculations below the defined
threshold for aridity (annual precipitation / annual potential
evapotranspiration); Default: 0.2}

\item{cut_runoff}{double. Exclude boundary calculations below the defined
runoff threshold; Default: 0 mm per year (no treshold)}

\item{with_groundwater_denit}{logical. Include global assumptions made on
groundwater denitrification losses. Defaults to TRUE ( = simulated leaching
is multiplied with 0.71 based on simulated denitrification losses in ground
water from Bouwman et al 2013)}

\item{prefix_monthly_output}{character. Provide a prefix if required for
monthly LPJmL output files, e.g. \code{"m"} for \code{"mdischarge.bin"} instead of
\code{"discharge.bin"} (default is \code{""})}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info). To be used for
time series analysis}
}
\description{
Calculate the PB status for the nitrogen boundary based on a scenario LPJmL
run and if \code{method == "braun2022_minusref"} a reference LPJmL run.
}
\examples{
\dontrun{
 calc_nitrogen_status(path_scenario, path_reference)
}

}
