% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_status.R
\name{calc_status}
\alias{calc_status}
\title{Calculate the planetary boundary status}
\usage{
calc_status(
  boundary,
  path_scenario,
  path_reference,
  time_span_scenario = c(1982, 2011),
  time_span_reference = NULL,
  ...
)
}
\arguments{
\item{boundary}{character vector, boundary for which status is calculated.
Available terrestrial boundaries are c("bluewater", "greenwater", "lsc",
"nitrogen")}

\item{path_scenario}{output directory (character string) of the scenario
LPJmL run where binary files (soon with metafiles) are written}

\item{path_reference}{output directory (character string) of the reference
LPJmL run where binary files (soon with metafiles) are written}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{...}{further arguments to be passed to each calc_* function}
}
\value{
list with data array for each \code{boundary}
}
\description{
Calculate the PB status for a defined planetary boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{ 
 boundary_status <- calc_status(
   boundary = 
   path_scenario = "./my_scenario/output",
   path_reference = "./my_reference/output")
}

}
