% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_water_deviations.R
\name{calc_water_deviations}
\alias{calc_water_deviations}
\title{Calculate water status based on deviations of a monthly scenario variable
from a corresponding monthly reference variable}
\usage{
calc_water_deviations(
  files_scenario,
  files_reference,
  spatial_scale = "subglobal",
  time_span_scenario = NULL,
  time_span_reference,
  approach = "porkka2024",
  thresholds = NULL,
  time_series_avg = NULL,
  config_args = list(),
  variable = "rootmoist"
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. All needed files are
provided in XXX. E.g.: list(leaching = "/temp/leaching.bin.json")}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the reference LPJmL run. All needed files are
provided in XXX. E.g.: list(leaching = "/temp/leaching.bin.json"). If not
needed for the applied approach, set to NULL.}

\item{spatial_scale}{character string indicating spatial scale;
"global" or "subglobal" for calculation of the share (\%) of total
global/basin area with deviations (either one value per year
(wang-erlandsson2022) or one value per year and month (porkka2024)); "grid"
not yet defined}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as character string, e.g. \code{as.character(1982:2011)} (default)}

\item{time_span_reference}{time span to be used for the reference run,
defined as a character string (e.g. \code{as.character(1901:1930)}).
Can differ in offset and length from \code{time_span_scenario}!
If \code{NULL} value of \code{time_span_scenario} is used}

\item{approach}{approach (character string) to be used , currently available
approach is \code{c("wang-erlandsson2022")} based on
\href{https://doi.org/10.1038/s43017-022-00287-8}{Wang-Erlandsson et al. 2022}
(referring only to the driest/wettest month of each year) or
\code{porkka2024} based on
\href{https://eartharxiv.org/repository/view/3438/}{Porkka et al. 2023}
(referring to each month of a year; default)}

\item{thresholds}{list with thresholds to be used to
define the safe, increasing risk and high risk zone,
For spatial_scale = "global" and "subglobal", this refers to the quantiles of
the global/basin area with deviations in the reference period. The default
is: c(holocene = 50, pb = 95, highrisk = NULL).
If set to NULL, the  default is taken from metric_files.yml
For highrisk, the value is currently hard-coded to 0.5
(following Richardson et al. 2023)}

\item{time_series_avg}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for calculation,
for \code{1} the whole time span is used to calculate a time
series.}

\item{config_args}{list of arguments to be passed on from the model
configuration.}

\item{variable}{character string with the name of the variable to be used
for the calculation of the water deviations. Default is "rootmoist"}
}
\description{
Calculate deviations (<q5 / >q95) for a monhtly variable in a scenario LPJmL
run as compared to a reference LPJmL run, either referring to global area
share with deviations (spatial_scale: global), or to number of months or
years with deviations (spatial resolution: cell). From this, calculate a
global or gridded PB status
}
