% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANPP.R
\name{calcHANPP}
\alias{calcHANPP}
\title{Calculate the ecosystem change metric gamma between 2 simulations/timesteps}
\usage{
calcHANPP(
  inFol_lu,
  inFol_pnv,
  startyr,
  stopyr,
  gridbased = T,
  pftbands = 11,
  cftbands = 32,
  p = 1,
  readPreviouslySavedData = F,
  saveDataFile = T,
  dataFile,
  ncells = 67420,
  fileType = "clm",
  headersize = 0,
  varnames
)
}
\arguments{
\item{inFol_lu}{folder of landuse scenario run}

\item{inFol_pnv}{folder of pnv reference run}

\item{startyr}{first year of simulations}

\item{stopyr}{last year of simulations}

\item{gridbased}{logical are pft outputs gridbased or pft-based?}

\item{pftbands}{number of natural plant functional types (== bands in fpc - 1)}

\item{cftbands}{number of crop functional types}

\item{p}{fraction of pasture band to take}

\item{readPreviouslySavedData}{flag whether to read previously saved data
instead of reading it in from output files (default F)}

\item{saveDataFile}{whether to save input data to file (default T)}

\item{dataFile}{file to save computed hanpp data to (default NULL)}

\item{ncells}{number of cells in lpjml grid}

\item{fileType}{type of output files - one of "clm", "nc", "meta" (default: "clm")}

\item{headersize}{headersize of the output files (default 0)}

\item{varnames}{data.frame with names of output files -- can be specified to account for variable file names (default NULL -- standard names are used)}
}
\value{
list data object containing arrays of ...
}
\description{
Function to calculate the ecosystem change metric gamma according
to Sykes (1999), Heyder (2011), and Ostberg (2015,2018).
This is a reformulated version in R, not producing 100\% similar values
than the C/bash version from Ostberg 2018, but following their methodology.
}
\examples{
\dontrun{
}
}
