% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_biomes.R, R/classify_biomes2.R
\name{classify_biomes}
\alias{classify_biomes}
\title{Classify biomes}
\usage{
classify_biomes(
  path_data,
  time_span = c(1982, 2011),
  vegc_proxy = TRUE,
  avg_nyear_args = list(),
  start_year = 1901
)

classify_biomes(
  path_data,
  time_span = c(1982, 2011),
  vegc_proxy = TRUE,
  avg_nyear_args = list(),
  start_year = 1901
)
}
\arguments{
\item{path_data}{output directory (character string) of the LPJmL run where
binary files (soon with metafiles) are written}

\item{time_span}{time span to be used, defined as an integer vector,
e.g. \code{1982:2011} (default)}

\item{vegc_proxy}{logical. Use vegetation carbon (7500 gC/m2) as a proxy
threshold to distinguish forests and savannahs}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info). To be used for
time series analysis}

\item{data}{list object. Contains all relevant data for computation of the
biome classes as yearly averages of monthly values.
fpc \link{dim=c(ncells,npfts+1)},pft_lai \link{dim=c(ncells,npfts+ncfts)},
lat, lon, vegc, temp \link{all dim=c(ncells)} e.g. list(lat = lat67420,
lon = lon67420, fpc = apply(fpc_scen, c(1,2), mean),
vegc = mean_state_scen\link{,8},temp = average_annual_temperature_1916)}

\item{readOutput}{read output from folder or use supplied data
(requires folder and timespan to be set)}

\item{folder}{to read outputs from}

\item{files}{to read as list list(grid="grid.bin",fpc="fpc.bin")}

\item{timespan}{as c(startyear,stopyear) to use for averaging outputs over}

\item{savannaProxy}{"vegc" or "natLAI". Use vegetation carbon or LAI in natural
vegetation as a proxy threshold to distinguish forests and savannahs}

\item{montaneArcticProxy}{"elevation" or "latitude". Use elevation or latitude
as a proxy threshold to distinguish arctic tundra and montane grassland}

\item{elevation_threshold}{threshold in m above which ArcticTundra is
classified as Montane Grassland, if montaneArcticProxy is set to
elevation - default: 1000}

\item{latitude_threshold}{threshold in degrees, south of which ArcticTundra is
classified as Montane Grassland, if montaneArcticProxy is set to
latitude - default: 55}

\item{lai_threshold}{threshold for "natLAI" proxy (default: 6 m2/m2)}

\item{vegc_threshold}{threshold for "vegc" proxy (default: 7500 gC/m2)}

\item{lpjGridInput}{path to lpjml grid input to be used for conversion
only required for nc input
(default: "/p/projects/lpjml/input/historical/input_VERSION2/grid.bin")}

\item{elevationInput}{path to lpjml elevation input to be used as proxy
(default: "/p/projects/lpjml/input/historical/input_VERSION2/elevation.bin")}

\item{lpjGridHeaderSize}{header size for lpjml grid input (default: 43)
only required for nc input}

\item{lpjCells}{number of grid cells in lpjml grid input (default: 67420)
only required for nc input}
}
\value{
list object containing biome_id (main biome per grid cell \link{dim=c(ncells)}),
and list of respective biome_names, biome_names_short, biomes_abbrv \link{dim=c(nbiomes)}

array with cell and second dimension if avg_nyear_args is specified
}
\description{
Classify biomes based on foliage protected cover (FPC) LPJmL output
and temperature output plus either vegetation carbon or pft_lai depending on
the savannaProxy option and elevation if montaneArcticProxy requires this

Classify biomes based on foliage protected cover (FPC) LPJmL output
}
\examples{
\dontrun{
classify_biomes(readOutput = T, timespan = c(1982,2011)
     folder = "/p/projects/open/Fabian/runs/Gamma/output/historic_gamma/",
     files = list(grid = "grid.bin", fpc = "fpc.bin", vegc = "vegc.bin",
     pft_lai = "pft_lai.bin", temp = "temp.bin"))
}

}
