% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_biomes2.R
\name{classify_biomes2}
\alias{classify_biomes2}
\title{Classify biomes}
\usage{
classify_biomes2(
  path_data,
  time_span = c(1982, 2011),
  vegc_proxy = TRUE,
  avg_nyear_args = list(),
  start_year = 1901
)
}
\arguments{
\item{path_data}{output directory (character string) of the LPJmL run where
binary files (soon with metafiles) are written}

\item{time_span}{time span to be used, defined as an integer vector,
e.g. \code{1982:2011} (default)}

\item{vegc_proxy}{logical. Use vegetation carbon (7500 gC/m2) as a proxy
threshold to distinguish forests and savannahs}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info). To be used for
time series analysis}
}
\value{
array with cell and second dimension if avg_nyear_args is specified
}
\description{
Classify biomes based on foliage protected cover (FPC) LPJmL output
}
