% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efr_aggregation.R
\name{efr_aggregation}
\alias{efr_aggregation}
\title{Aggregate discharge composition as EFRs and PBs from grid cell to global}
\usage{
efr_aggregation(
  endcell,
  routing,
  cellindex,
  efr,
  discharge,
  remotes = F,
  flowseason = F
)
}
\arguments{
\item{endcell}{array containing the index of each cells final drainage cell}

\item{routing}{array containing the index of each cells next drainage cell}

\item{cellindex}{array containing the rank of each cell in the drainage
network (1: final drainage cell, 2: draining directly into final
drainage cell, ...)}

\item{efr}{list}

\item{discharge}{list}

\item{remotes}{list}

\item{flowseason}{list}
}
\value{
global array of aggregated EFR and PB
}
\description{
Aggregate discharge composition as EFRs and PBs from grid cell to global
}
\examples{
\dontrun{
}
}
