% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenwater_status.R
\name{greenwater_status}
\alias{greenwater_status}
\title{Status calculation of the greenwater boundary}
\usage{
greenwater_status(
  files_scenario,
  files_reference,
  spatial_scale = "global",
  time_span_scenario = as.character(1982:2011),
  time_span_reference = time_span_scenario,
  approach = "wang-erlandsson2022",
  time_resolution = NULL,
  config_args = list(),
  thresholds = NULL
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. Handled automatically via
\code{\link[=calc_status]{calc_status()}}.}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the files_reference LPJmL run. Handled automatically via
\code{\link[=calc_status]{calc_status()}}.}

\item{spatial_scale}{character string indicating spatial resolution
either "grid", "subglobal" or "global" for calculation of the share (\%)
of total global area with deviations}

\item{time_span_scenario}{time span to use output from the scenario run,
e.g. \code{1982:2011}.}

\item{time_span_reference}{time span use output from the reference run,
e.g. \code{1901:1930}.}

\item{approach}{approach (character string) to be used , currently available
approach is \code{c("wang-erlandsson2022")} based on
\href{https://doi.org/10.1038/s43017-022-00287-8}{Wang-Erlandsson et al. 2022}
(referring only to the driest/wettest month of each year) or
\code{porkka2024} based on
\href{https://eartharxiv.org/repository/view/3438/}{Porkka et al. 2023}
(referring to each month of a year)}

\item{time_resolution}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for one status
calculation, for \code{1} the whole time span is used to calculate a status time
series.}

\item{config_args}{list of arguments to be passed on from the model
configuration.}

\item{thresholds}{named character string with thresholds to be used to
define the safe, increasing risk and high risk zone,
e.g. c(holocene = 0.5, pb = 0.95, highrisk = 0.99).
For spatial resolution = "grid", this refers to the p value
(significance level of increases in deviations) with the default:
c(holocene = 1, pb = 0.05, highrisk = 0.01).
For spatial resolution = "global", this refers to the quantiles of
the global area with deviations in the reference period. The dafault
for global resolution is: c(holocene = 0.5, pb = 0.95,
highrisk = 0.99).
If set to NULL, the respective default is taken (see above; matching
the spatial_scale, defined in metric_files.yml).}
}
\value{
Object of class \code{control_variable} with the boundary status of the
greenwater boundary.
}
\description{
Planetary Boundary status calculation of the greenwater boundary based on
rootmoisture in a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
boundary_status <- calc_status(
  boundary = "greenwater",
  config_scenario = "path/to/config_scenario.json",
  config_reference = "path/to/config_reference.json",
  spatial_scale = "global",
  time_span_scenario = 1901:2019,
  time_span_reference = 1901:1930,
  approach = "porkka2024"
)
}

}
