% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{listOutputs}
\alias{listOutputs}
\title{Returns LPJmL outputs required for given metric}
\usage{
listOutputs(metric = "both", withNitrogen = TRUE)
}
\arguments{
\item{metric}{string: which metric? "meco", "mcol", "biome", or "all" (default)}

\item{withNitrogen}{logical: include nitrogen outputs? default: TRUE}
}
\value{
List of strings with LPJmL variable names
}
\description{
Function to return a list of strings with the LPJmL output names required for
the computation of the metrics M-COL, M-ECO, or the biome classification
}
\examples{
\dontrun{

}
}
