% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsc_status.R
\name{lsc_status}
\alias{lsc_status}
\title{Status calculation of the land-system change boundary}
\usage{
lsc_status(
  files_scenario,
  files_reference,
  spatial_scale = "subglobal",
  time_span_scenario = as.character(1982:2011),
  time_span_reference = time_span_scenario,
  approach = "steffen2015",
  time_series_avg = NULL,
  config_args = list(),
  thresholds = NULL,
  eurasia = TRUE,
  ...
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. Handled automatically via
\code{\link[=calc_status]{calc_status()}}.}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the files_reference LPJmL run. Handled automatically via
\code{\link[=calc_status]{calc_status()}}.}

\item{spatial_scale}{character. Spatial resolution, available options
are \code{"subglobal"} (at the biome level, default), \code{"global"} and \code{"grid"}}

\item{time_span_scenario}{time span to use output from the scenario run,
e.g. \code{1982:2011}.}

\item{time_span_reference}{time span use output from the reference run,
e.g. \code{1901:1930}.}

\item{approach}{approach (character string) to be used , currently available
approach is \code{"steffen2015"}}

\item{time_series_avg}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for one status
calculation, for \code{1} the whole time span is used to calculate a status time
series.}

\item{config_args}{list of arguments to be passed on from the model
configuration.}

\item{thresholds}{list with deforestation thresholds for defining safe,
increasing risk and high risk zone. Default based on Steffen et al. 2015 if
thresholds set to NULL (https://doi.org/10.1126/science.1259855):
for gridded and biome scale application:
`list(pb = list(temperate = 0.5, tropical = 0.15, boreal = 0.15),
highrisk = list(temperate = 0.7, tropical = 0.4, boreal = 0.4))
for global scale application:
list(holocene = 0, pb = 0.25, highrisk = 0.46)
pb = threshold between safe zone and increasing risk zone
(e.g. 50\% for boreal forest with default value)
highrisk = threshold between increasing risk and high risk zone}

\item{eurasia}{logical. If \code{spatial_scale} = \code{"subglobal"} merge continents
Europe and Asia to avoid arbitrary biome cut at europe/asia border.
Defaults to \code{TRUE}}

\item{...}{arguments forwarded to \code{\link{classify_biomes}}}
}
\value{
Object of class \code{control_variable} with the boundary status of the
lsc boundary.
}
\description{
Planetary Boundary status calculation of the LSC (land-system change) boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
boundary_status <- calc_status(
  boundary = "lsc",
  config_scenario = "path/to/config_scenario.json",
  config_reference = "path/to/config_reference.json",
  spatial_scale = "global",
  time_span_scenario = 1901:2019,
  time_span_reference = 1901:1930
)
}

}
