% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minkowski_offset.R
\name{minkowski_offset}
\alias{minkowski_offset}
\title{Offset polygon using minkowski sum with a disc}
\usage{
minkowski_offset(polygon, offset, arc_segments = 50, eps = 1e-05)
}
\arguments{
\item{polygon}{A \code{polyclid_polygon} vector. If shorter than \code{offset} it will
be recycled to the length of \code{offset}}

\item{offset}{An \code{euclid_exact_numeric} or numeric vector. If shorter than
\code{polygon} it will be recycled to the length of \code{polygon}}

\item{arc_segments}{The number of segments used to draw a full circle}

\item{eps}{The error bound of the approximation}
}
\value{
A \code{polyclid_polygon_set} vector
}
\description{
The minkowski sum of two polygons is the polygon you get when you trace one
polygon around the boundary of the other and adding the covered area to the
boundary. As such it provides an offset with rounded corners for obtuse
angles. The algorithm provided here is an approximation with the allowed
error bound being controlled by the \code{eps} argument. Further, the introduced
corner arcs are estimations of the disc as they are converted to line
segments.
}
\examples{
poly <- polyclid::denmark()

plot(poly, col = "grey")

# Negative offset (inset)
ins <- minkowski_offset(poly, -0.05, arc_segments = 10)
euclid_plot(ins, lty = 2)

# Positive offset
off <- minkowski_offset(poly, 0.1, arc_segments = 10)
euclid_plot(off, lty = 3)

}
\seealso{
Other polygon offsetting: 
\code{\link{skeleton_offset}()}

Other minkowski sum functions: 
\code{\link{minkowski_sum}()}
}
\concept{minkowski sum functions}
\concept{polygon offsetting}
