% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minkowski_sum.R
\name{minkowski_sum}
\alias{minkowski_sum}
\title{Calculate the minkowski sum of two polygons}
\usage{
minkowski_sum(P, Q)
}
\arguments{
\item{P}{A \code{polyclid_polygon} vector that should be expanded}

\item{Q}{A \code{polyclid_polygon} vector to trace around \code{P}}
}
\value{
A \code{polyclid_polygon} vector
}
\description{
The minkowski sum is the point-wise sum of two polygons. Practically it means
that you trace one polygon around the boundary of the other, expanding the
polygon into the covered area.
}
\examples{

poly <- polyclid::denmark()

# Expand Jutland by tracing the Island of Bornhold around it

# First we should center Bornholm on the origin.
bornholm <- transform(poly[7], affine_translate(-vec(centroid(poly[7]))))
plot(minkowski_sum(poly[1], bornholm))

}
\seealso{
Other minkowski sum functions: 
\code{\link{minkowski_offset}()}
}
\concept{minkowski sum functions}
