% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_status.R
\name{nitrogen_status}
\alias{nitrogen_status}
\title{Status calculation of the nitrogen boundary}
\usage{
nitrogen_status(
  files_scenario,
  files_reference,
  spatial_scale = "grid",
  time_span_scenario = 1982:2011,
  time_span_reference = time_span_scenario,
  approach = "braun2022",
  time_resolution = NULL,
  config_args = list(),
  thresholds = NULL,
  cut_arid = 0.2,
  cut_runoff = 0,
  with_groundwater_denit = TRUE
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. Handled automatically via
\code{\link[=calc_status]{calc_status()}}.}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the files_reference LPJmL run. Handled automatically via
\code{\link[=calc_status]{calc_status()}}.}

\item{spatial_scale}{character. Spatial resolution, available options
are \code{"global"} and \code{"grid"}}

\item{time_span_scenario}{time span to use output from the scenario run,
e.g. \code{1982:2011}.}

\item{time_span_reference}{time span use output from the reference run,
e.g. \code{1901:1930}.}

\item{approach}{(character string) to be used , currently available
approach is \code{"braun2022"} based on unpublished suggestion by Johanna Braun.
Second approach option is \code{"braun2022_minusref"} to subtract reference run
output}

\item{time_resolution}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for one status
calculation, for \code{1} the whole time span is used to calculate a status time
series.}

\item{config_args}{list of arguments to be passed on from the model
configuration.}

\item{thresholds}{list with highrisk and pb threshold for N concentration
(mg N/l) in runoff to surface water
Default: highrisk = 5, pb = 2
(based on Schulte-Uebbing et al. 2022,
https://doi.org/10.1038/s41586-022-05158-2:
"we used a threshold for N concentration in run-off to surface water. This
threshold was set to 5.0 mgN/l, based on the assumption that on average 50\%
of N entering surface water is removed through retention and sedimentation"))}

\item{cut_arid}{double. Exclude boundary calculations below the defined
threshold for aridity (annual precipitation / annual potential
evapotranspiration); Default: 0.2}

\item{cut_runoff}{double. Exclude boundary calculations below the defined
runoff threshold; Default: 0 mm per year (no treshold)}

\item{with_groundwater_denit}{logical. Include global assumptions made on
groundwater denitrification losses. Defaults to TRUE ( = simulated leaching
is multiplied with 0.71 based on simulated denitrification losses in ground
water from Bouwman et al 2013)}
}
\value{
Object of class \code{control_variable} with the boundary status of the
nitrogen boundary.
}
\description{
Planetary Boundary status calculation of the the nitrogen boundary based on
a scenario LPJmL run and if \code{approach == "braun2022_minusref"} a reference
LPJmL run.
}
\examples{
\dontrun{
boundary_status <- calc_status(
  boundary = "nitrogen",
  config_scenario = "path/to/config_scenario.json",
  config_reference = "path/to/config_reference.json",
  spatial_scale = "global",
  time_span_scenario = 1901:2019,
  time_span_reference = 1901:1930
)
}

}
