% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomes.R
\name{plot_biomes}
\alias{plot_biomes}
\title{Plot global distribution of lpjml simulated biomes}
\usage{
plot_biomes(
  biome_data,
  file_name = NULL,
  to_robinson = TRUE,
  bg_col = "white",
  grid_path = NULL
)
}
\arguments{
\item{biome_data}{output (list) from classify_biomes()}

\item{file_name}{directory for saving the plot (character string)}

\item{to_robinson}{logical to define if robinson projection should be used
for plotting}

\item{bg_col}{character, specify background possible (\code{NA} for transparent)#}

\item{grid_path}{character string providing the path to a grid file}
}
\description{
Plots a map with the biome distribution as derived from a lpjml run based
on the "classify_biomes" function
}
\examples{
\dontrun{
plot_biomes(
  biome_data = biomes,
  file_name = "/p/projects/open/Johanna/R/biomes.pfd"
  grid_path = ".grid.bin.json"
)
}

}
