% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boundaries.R
\name{plot_boundaries}
\alias{plot_boundaries}
\title{Plot polar boundaries plot including time series of boundaries}
\usage{
plot_boundaries(x, normalization = "increasing risk")
}
\arguments{
\item{x}{boundaries object}

\item{normalization}{see \code{\link[=as_risk_level]{as_risk_level()}}}
}
\description{
Plot time series of boundaries into iconic polar boundaries plot only
focussing on the terrestrial boundaries (half-circle). Wedges are
scaled and normalized according to each boundary.
}
\examples{
\dontrun{
data_path_ma <- (
  "/p/projects/open/Johanna/boundaries/R/r_out/r_data/global_timeseries_avg_10_porkka.RData"
)

data <- load(data_path_ma) |>
  get()

ggplot2::ggsave(
  "test_boundaries_plot.png",
  plot_boundaries(data),
  width = 16,
  height = 9,
  dpi = 300
)
}

}
