% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status.R
\name{plot_status}
\alias{plot_status}
\title{Plot the global status of planetary boundaries}
\usage{
plot_status(
  x,
  file_name = NULL,
  colors = c(`safe zone` = "#74bca9e1", `increasing risk` = "#f6ee0f", `high risk` =
    "#e23a50"),
  bg_col = "white",
  to_robinson = TRUE,
  legend = TRUE
)
}
\arguments{
\item{x}{list with one to four LPJmL vectors (length: ncells) with
the outputs from the calc_* functions for one point in time.
Given element names within the list will be displayed underneath the
respective map. Order in the list determines the plotting order (by row).}

\item{file_name}{character string providing file name (including directory
and file extension). Defaults to NULL (plotting to screen)}

\item{colors}{definition of colors for plotting of the safe, increasing
risk and high risk zone (3 elements, do not change the order). Colors used
for previous plots: "#008B0099" (green), "#FFEB00", "brown3"}

\item{bg_col}{character, specify background possible (\code{NA} for transparent)}

\item{to_robinson}{logical to define if robinson projection should be used
for plotting}

\item{legend}{logical, specify whether a legend should be plotted}
}
\description{
Plot global map(s) with the status of 1-4 planetary boundaries for
a scenario LPJmL run and derived planetary boundary statuses
}
\examples{
\dontrun{
 plot_status(file_name = "./my_boundary_status.png",
                x = list("land system change" = lsc_status,
                                   "nitrogen" = nitrogen_status,
                                   "bluewater" = water_status),
 legend = FALSE
 bg_col = NA)
}

}
