% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_global.R
\name{plot_status_global}
\alias{plot_status_global}
\title{Plot the global status of planetary boundaries}
\usage{
plot_status_global(
  x,
  file_name = NULL,
  colors = c(`safe zone` = "#e3f5ec", `increasing risk` = "#fcf8d8", `high risk` =
    "#f7e4dd"),
  all_in_one = FALSE,
  ncol = 2
)
}
\arguments{
\item{x}{list with global output from calc_status}

\item{file_name}{character string providing file name (including directory
and file extension). Defaults to NULL (plotting to screen)}

\item{colors}{named vector for definition of colors for plotting of
the safe, increasingnrisk and high risk zone}

\item{all_in_one}{boolean, if TRUE, all PB stati will be normalized to 0
(holocene value) to 1 (pb) and plotted in one panel}

\item{ncol}{number of plot columns (only relevant if more than one pb is
plotted and all_in_one = FALSE)}
}
\description{
Plot line plots with the PB status over time  for
a scenario LPJmL run and derived planetary boundary statuses
}
\examples{
\dontrun{
 plot_status(file_name = "./my_boundary_status.png",
                x = list("land system change" = lsc_status,
                                   "nitrogen" = nitrogen_status,
                                   "bluewater" = water_status),
 legend = FALSE
 bg_col = NA)
}

}
