% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_maps.R
\name{plot_status_maps}
\alias{plot_status_maps}
\title{Plot the global status of planetary boundaries}
\usage{
plot_status_maps(
  x,
  filename = NULL,
  add_legend = TRUE,
  projection = "+proj=robin",
  ncol = 2,
  grid_path = NULL
)
}
\arguments{
\item{x}{output object from calc_* with the status of the
control variable for one point in time, incl. pb thresholds as attribute}

\item{filename}{character string providing file name (including directory
and file extension). Defaults to NULL (plotting to screen)}

\item{add_legend}{logical, specify whether a add_legend should be plotted}

\item{projection}{character string defining the projection, default set to
"+proj=robin"}

\item{ncol}{integer, number of columns in the plot, default set to 2}

\item{grid_path}{character string providing the path to a grid file}
}
\description{
Plot global map(s) with the status of planetary boundaries for
a scenario LPJmL run and derived planetary boundary statuses
}
\examples{
\dontrun{
 plot_status_maps(
  filename = "./my_boundary_status.png",
  x = calc_output
  add_legend = FALSE
  grid_path = "/path/to/gridfile.bin.json"
)
}

}
