% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify_poly}
\alias{simplify_poly}
\title{Simplify polygons and polylines}
\usage{
simplify_poly(
  poly,
  cost = "squared",
  stop = "below count ratio",
  cost_ratio = 0.5,
  stop_threshold = 0.5
)
}
\arguments{
\item{poly}{A \code{polyclid_polygon} or \code{polyclid_polyline}}

\item{cost}{The cost function to use. See details.}

\item{stop}{The stop condition to use for terminating the simplification. See
details.}

\item{cost_ratio}{The ratio to use if \verb{cost = "hybrid squared}}

\item{stop_threshold}{The threshold to use with the stop condition}
}
\value{
A vector of the same type as \code{poly}
}
\description{
This function allows you to reduce the number of vertices used to encode the
geometry while attempting to keep the geometry as close to the original as
possible. You can choose between different cost functions and stop
conditions in order to fine tune the procedure.
}
\examples{
poly <- polyclid::denmark()

plot(poly)
poly_s <- simplify_poly(poly)

# Number of original vertices
sum(cardinality(poly))

# Number after simplification
sum(cardinality(poly_s))

# Close to same visual
plot(poly_s)

# Plot Jutland using only 20 vertices
plot(simplify_poly(poly[1], stop = "below count", stop_threshold = 20))
euclid_plot(poly[1], lty = 2, border = "red")

}
