% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_limit.R
\name{skeleton_limit}
\alias{skeleton_limit}
\alias{skeleton_limit_location}
\title{Limit of straight skeletons}
\usage{
skeleton_limit(polygon)

skeleton_limit_location(polygon)
}
\arguments{
\item{polygon}{A \code{polyclid_polygon} vector}
}
\value{
\code{skeleton_limit} returns a \code{euclid_exact_numeric} vector and
\code{skeleton_limit_location} returns a \code{polyclid_polyline_set} vector
}
\description{
The limit of a straight skeleton is the distance at which the inner offset
disappears completely. You can get the geometry left at the limit as a
polyline set consisting of isolated vertices and/or segments.
}
\examples{

poly <- polyclid::denmark()[9]
# Complex polygons will often have a single limit point

plot(poly)
euclid_plot(vert(skeleton_limit_location(poly)))

# But certain geometries will result in segments rather than lines
poly <- polyclid::polygon(
  c(1, 1, 2, 2, -2, -2, -1, -1),
  c(-2, 0, 0, 2, 2, 0, 0, -2)
)
euclid_plot(skeleton_limit_location(poly))

# You can get the distance from the limit location to the boundary
skeleton_limit(poly)

}
\seealso{
Other straight skeleton functions: 
\code{\link{skeleton_interior}()},
\code{\link{skeleton_offset}()}
}
\concept{straight skeleton functions}
