% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_offset.R
\name{skeleton_offset}
\alias{skeleton_offset}
\title{Offset a polygon based on it's straight skeleton}
\usage{
skeleton_offset(polygon, offset)
}
\arguments{
\item{polygon}{A \code{polyclid_polygon} vector. If shorter than \code{offset} it will
be recycled to the length of \code{offset}}

\item{offset}{An \code{euclid_exact_numeric} or numeric vector. If shorter than
\code{polygon} it will be recycled to the length of \code{polygon}}
}
\value{
A \code{polyclid_polygon_set} vector
}
\description{
A straight skeleton is the structure made up of the angular bisectors of each
vertex in a polygon (see \code{\link[=skeleton_interior]{skeleton_interior()}} for how to get the skeleton).
It can be used to offset polygons either inward or outward. This offsetting
technique will keep the corners of the polygon \emph{as-is}.
}
\examples{
poly <- polyclid::denmark()

plot(poly, col = "grey")

# Negative offset (inset)
ins <- skeleton_offset(poly, -0.05)
euclid_plot(ins, lty = 2)

# Positive offset
off <- skeleton_offset(poly, 0.1)
euclid_plot(off, lty = 3)

# As can be seen, small sharp angles can become severely exaggerated with
# this offset type
plot(skeleton_offset(poly[9], c(0.01, 0.02, 0.04, 0.06, 0.1)))

}
\seealso{
Other polygon offsetting: 
\code{\link{minkowski_offset}()}

Other straight skeleton functions: 
\code{\link{skeleton_interior}()},
\code{\link{skeleton_limit}()}
}
\concept{polygon offsetting}
\concept{straight skeleton functions}
