% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_legend.R
\name{status_legend}
\alias{status_legend}
\title{Plot the legend for the normalized colors of PB statuses}
\usage{
status_legend(filename = NULL, fontsize = 3)
}
\arguments{
\item{filename}{character string providing file name (including directory
and file extension). Defaults to NULL (return
plot object for further adaptation)}

\item{fontsize}{numeric specifying the size of the font to be used for
legend labels. Default set to 3.}
}
\description{
Plot a legend for the colors of PB statuses, normalized based on the size of
the increasing risk, for globally aggregated plots,
or spatially distributed maps
}
\examples{
\dontrun{
status_legend(
  filename = "./mylegend.png",
)
}

}
