% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{validation_table}
\alias{validation_table}
\title{Validate simulated global PB-relevant variables against literature
Calculate a table with global modelled vs literature values for key variables
relevant to planetary boundaries}
\usage{
validation_table(
  config_scenario,
  config_reference,
  time_span_scenario,
  time_span_reference,
  path_baseline,
  table_path,
  gridbased = TRUE,
  ...
)
}
\arguments{
\item{config_scenario}{character string. File path to the LPjmL configuration
file (json) of the scenario run. The configuration file contains the
information about the LPJmL run, e.g. the output directory}

\item{config_reference}{character string. See config_scenario. For the
reference run}

\item{time_span_scenario}{time span to be used for the scenario run and
parallel PNV run, defined as a character string,
e.g. \code{as.character(1982:2011)}}

\item{time_span_reference}{time span to be used for the reference run, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{path_baseline}{character string for path to outputs
for the baseline run, file names are taken from files scenario}

\item{table_path}{character string for file  path to save the output,  .csv file}

\item{gridbased}{logical, set to TRUE for gridbased pft outputs}

\item{...}{arguments to be passed to \link{calc_status}}
}
\value{
table with comparison between lpjml values and literature ranges
}
\description{
Validate simulated global PB-relevant variables against literature
Calculate a table with global modelled vs literature values for key variables
relevant to planetary boundaries
}
\examples{
\dontrun{
validation <- validation_table(
  table_path = "./my_path/table.csv",
  timeframe = as.character(2010:2017)
)
}

}
