% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_time.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Calculate averages (mean) for defined window sizes}
\usage{
aggregate_time(x, time_resolution = NULL, time_repeat = NULL)
}
\arguments{
\item{x}{LPJmL output array with \code{dim(x)=c(cell, month, year)}}

\item{time_resolution}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for
\code{spatial_scale = c("grid", "subglobal")} or the whole time span is used for
\code{spatial_scale = "global"}.}

\item{time_repeat}{integer, if supplied (default NULL), it defines a
length of years to be replicated. Only if time_resolution is not supplied.}
}
\value{
array with same amount of cells and months as x if time_resolution is
supplied. If time_repeat is supplied, the array has the same amount of
cells and months as x but the amount of years is multiplied by
time_repeat.
}
\description{
Define window sizes (time_resolution) to be used to calculate moving averages
(mean). If time_resolution is not supplied, the function calculates the mean
over all years. If time_repeat is supplied, the function replicates the
mean values for the defined amount of years.
}
