% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_time.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Calculate averages (mean) for defined window sizes}
\usage{
aggregate_time(x, nyear_window = NULL, nyear_replicate = NULL)
}
\arguments{
\item{x}{LPJmL output array with \code{dim(x)=c(cell, month, year)}}

\item{nyear_window}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for
\code{spatial_scale = c("grid", "subglobal")} or the whole time span is used for
\code{spatial_scale = "global"}.}

\item{nyear_replicate}{integer, if supplied (default NULL), it defines a
length of years to be replicated. Only if nyear_window is not supplied.}
}
\value{
array with same amount of cells and months as x if nyear_window is
supplied. If nyear_replicate is supplied, the array has the same amount of
cells and months as x but the amount of years is multiplied by
nyear_replicate.
}
\description{
Define window sizes (nyear_window) to be used to calculate moving averages
(mean). If nyear_window is not supplied, the function calculates the mean
over all years. If nyear_replicate is supplied, the function replicates the
mean values for the defined amount of years.
}
