% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries-package.R
\docType{package}
\name{boundaries-package}
\alias{boundaries}
\alias{boundaries-package}
\title{boundaries: Planetary Boundary Status with LPJmL simulations}
\description{
A systematic approach to quantify the status of the terrestrial planetary boundaries based on the Dynamic Global Vegetation Model (DGVM) Lund-Potsdam-Jena managed Land (LPJmL) hosted at the Potsdam Institute for Climate Impact Research (PIK). The supported planetary boundaries are "biosphere integrity", "land-system change", "bluewater", "greenwater" and "nitrogen flows".
}
\author{
\strong{Maintainer}: Jannes Breier \email{jannesbr@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9055-6904}{ORCID})

Authors:
\itemize{
  \item Fabian Stenzel \email{stenzel@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5109-0048}{ORCID})
  \item Johanna Braun \email{braun@pik-potsdam.de} (\href{https://orcid.org/0000-0002-0000-0000}{ORCID})
}

}
\keyword{internal}
