% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries-package.R
\docType{package}
\name{boundaries-package}
\alias{boundaries}
\alias{boundaries-package}
\title{boundaries: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
