% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bluewater_status.R
\name{calc_bluewater_status}
\alias{calc_bluewater_status}
\title{Calculate the planetary boundary status for the bluewater boundary}
\usage{
calc_bluewater_status(
  files_scenario,
  files_reference,
  time_span_scenario = NULL,
  time_span_reference = time_span_scenario,
  method = "gerten2020",
  cut_min = 0.0864,
  time_aggregation_args = list(),
  spatial_scale,
  thresholds = NULL
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. All needed files are
provided in metric_files.yaml.
E.g.: list(leaching = "/temp/leaching.bin.json")}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the reference LPJmL run. All needed files are
provided in metric_files.yml.
E.g.: list(leaching = "/temp/leaching.bin.json").}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as character string}

\item{time_span_reference}{time span to be used for the scenario run, defined
as character string, e.g. \code{as.character(1901:1930)}. Can differ in offset and
length from \code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is
used}

\item{method}{method (character string) to be used , currently available
method is \code{"gerten2020"} based on
\href{https://doi.org/10.1038/s41893-019-0465-1}{Gerten et al. 2020}
for spatial_scale = "grid" and
"wang_erlandsson2022" as well as "porkka2023" for
spatial_scale = "global" or "subglobal"}

\item{cut_min}{double. Exclude boundary calculations for discharge < cut_min
and dismiss EFR transgresssions if < cut_min for "gerten2020" method,
Default: 0.0864 hm3/day (=1 m3/s)}

\item{time_aggregation_args}{list of arguments to be passed to
\link[boundaries]{aggregate_time} (see for more info). To be used for
time series analysis}

\item{spatial_scale}{character string indicating spatial resolution
options: "global", "subglobal", "grid";
for "grid" the method "gerten2020" is applicable based on EFR calculations;
for "global"/"subglobal" the share (\%) of total global/basin area with
deviations is calculated}

\item{thresholds}{named character string with thresholds to be used to
define the safe, increasing risk and high risk zone, the method and scale
specific default thresholds are defined in metric_files.yml are are applied
if thresholds are set to NULL.}
}
\value{
todo: describe returned object
}
\description{
Calculate the PB status for the bluewater (former freshwater) boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
 calc_bluewater_status(files_scenario, files_reference)
}

}
