% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bluewater_status.R
\name{calc_bluewater_status}
\alias{calc_bluewater_status}
\title{Calculate the planetary boundary status for the bluewater boundary}
\usage{
calc_bluewater_status(
  path_scenario,
  path_reference,
  time_span_scenario = c(1982, 2011),
  time_span_reference = NULL,
  method = "gerten2020",
  temporal_resolution = "annual",
  cut_min = 0.0864,
  avg_nyear_args = list(),
  start_year = 1901
)
}
\arguments{
\item{path_scenario}{output directory (character string) of the scenario
LPJmL run where binary files (soon with metafiles) are written}

\item{path_reference}{output directory (character string) of the reference
LPJmL run where binary files (soon with metafiles) are written}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{method}{method (character string) to be used , currently available
method is \code{c("gerten2020")} based on
\href{https://doi.org/10.1038/s41893-019-0465-1}{Gerten et al. 2020}.}

\item{temporal_resolution}{character. Temporal resolution, available options
are \code{"annual"} (default) and \code{"monthly"}.}

\item{cut_min}{double. Exclude boundary calculations for Q < cut_min}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info). To be used for
time series analysis.}
}
\description{
Calculate the PB status for the bluewater (former freshwater) boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
 calc_bluewater(path_scenario, path_reference)
}

}
