% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_efrs.R
\name{calc_efrs}
\alias{calc_efrs}
\title{Calculate environmental flow requirements (EFRs)}
\usage{
calc_efrs(x, method = "vmf", avg_nyear_args = list())
}
\arguments{
\item{x}{discharge array with \code{dim(x)=c(cell, month, year)}}

\item{method}{EFR method to be used , available methods are \code{c("vmf", "vmf_min", "vmf_max", "q90q50")} based on
\href{https://doi.org/10.5194/hess-18-5041-2014}{Pastor et al. 2014}
as well as \code{"steffen2015"}, a modified version of vmf by
\href{https://doi.org/10.1126/science.1259855}{Steffen et al. 2015}}

\item{avg_nyear_args}{list of arguments to be passed to
\link[boundaries]{average_nyear_window} (see for more info). To be used for
time series analysis.}
}
\value{
EFRs with same unit as \code{x} (discharge), with \code{dim(x)=c(ncells, 12)}
or \code{dim(EFRs)=c(ncells, 12, dim(x)[3] / nyear_avg)} if nyear_avg is defined
}
\description{
Calculate environmental flow requirements (EFRs) based on the number of years
of \code{dim(x)[3]} or specify a nyear_avg calculate the EFRs for each bin in
\code{dim(x)[3]}.
}
\examples{
\dontrun{
# basic example
efrs1 <- calcEFRs(discharge_30y = discharge, method = "vmf")

dim(efrs1)
# c(67420, 12)

# example for using a 30 year average bin for a 90 year discharge and
#  interpolate between 3 windows afterwards to return 90 years (interpolated)
efrs2 <- calcEFRs(discharge_90y = discharge,
                  method="vmf",
                  avg_nyear_args = list(nyear_avg = 30, interpolate = TRUE))

dim(efrs2)
# c(67420, 12, 90)
# if interpolate == FALSE  dim(efrs2) returns c(67420, 12, 3)

# example for using a 1 year (no average) bin for a 100 year discharge
efrs3 <- calcEFRs(discharge_100y = discharge,
                  method = "vmfmin",
                  avg_nyear_args = list(nyear_avg = 1))

dim(efrs3)
# c(67420, 12, 100)
}
}
