% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_greenwater_status.R
\name{calc_greenwater_status}
\alias{calc_greenwater_status}
\title{Calculate the planetary boundary status for the greenwater boundary}
\usage{
calc_greenwater_status(
  files_scenario,
  files_reference,
  spatial_scale = "global",
  time_span_scenario = as.character(1982:2011),
  time_span_reference = time_span_scenario,
  approach = "wang-erlandsson2022",
  nyear_window = NULL,
  config_args = list(),
  thresholds = NULL
)
}
\arguments{
\item{files_scenario}{list with variable names and corresponding file paths
(character string) of the scenario LPJmL run. All needed files are
provided in metric_files.yml}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the reference LPJmL run. All needed files are
provided in metric_files.yml}

\item{spatial_scale}{character string indicating spatial resolution
either "grid", "subglobal" or "global" for calculation of the share (\%)
of total global area with deviations}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as a character string, e.g. \code{as.character(1982:2011)} (default)}

\item{time_span_reference}{time span to be used for the reference run,
defined as a character string (e.g. \code{as.character(1901:1930)}).
Can differ in offset and length from \code{time_span_scenario}!
If \code{NULL} value of \code{time_span_scenario} is used}

\item{approach}{approach (character string) to be used , currently available
approach is \code{c("wang-erlandsson2022")} based on
\href{https://doi.org/10.1038/s43017-022-00287-8}{Wang-Erlandsson et al. 2022}
(referring only to the driest/wettest month of each year) or
\code{porkka2024} based on
\href{https://eartharxiv.org/repository/view/3438/}{Porkka et al. 2023}
(referring to each month of a year)}

\item{nyear_window}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for one status
calculation, for \code{1} the whole time span is used to calculate a status time
series.}

\item{config_args}{list of arguments to be passed on from the model
configuration.}

\item{thresholds}{named character string with thresholds to be used to
define the safe, increasing risk and high risk zone,
e.g. c(holocene = 0.5, pb = 0.95, highrisk = 0.99).
For spatial resolution = "grid", this refers to the p value
(significance level of increases in deviations) with the default:
c(holocene = 1, pb = 0.05, highrisk = 0.01).
For spatial resolution = "global", this refers to the quantiles of
the global area with deviations in the reference period. The dafault
for global resolution is: c(holocene = 0.5, pb = 0.95,
highrisk = 0.99).
If set to NULL, the respective default is taken (see above; matching
the spatial_scale, defined in metric_files.yml).}
}
\description{
Calculate the PB status for the greenwater (former freshwater) boundary based
on rootmoisture in a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
 calc_greenwater_status(files_scenario, files_reference,
                time_span_reference, spatial_scale = "grid")
}

}
