% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lsc_status.R
\name{calc_lsc_status}
\alias{calc_lsc_status}
\title{Calculate the planetary boundary status for the land-system change boundary}
\usage{
calc_lsc_status(
  path_scenario,
  path_reference,
  time_span_scenario = c(1982, 2011),
  time_span_reference = NULL,
  spatial_resolution = "biome",
  eurasia = TRUE,
  biome_classification = NULL,
  vegc_proxy = FALSE,
  avg_nyear_args = list(),
  start_year = 1901
)
}
\arguments{
\item{path_scenario}{output directory (character string) of the scenario
LPJmL run where binary files (soon with metafiles) are written}

\item{path_reference}{output directory (character string) of the reference
LPJmL run where binary files (soon with metafiles) are written}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{spatial_resolution}{character. Spatial resolution, available options
are \code{"biome"} (default) and \code{"grid"}}

\item{eurasia}{logical. If \code{spatial_resolution} = \code{"biome"} merge continents
Europe and Asia to avoid arbitrary biome cut at europe/asia border. Defaults
to \code{TRUE}}

\item{biome_classification}{default to NULL, if biomes are to be calculated
within the function based on classify_biomes. Can be set to array
(dim = c(ncell, 1)) with already defined biomes.}

\item{vegc_proxy}{logical. If \code{TRUE} vegetation carbon (vegc) threshold of
7500 gC/m² is used to distinguish between forest and savannahs}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info). To be used for
time series analysis}
}
\description{
Calculate the PB status for the LSC (land-system change) boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{
\dontrun{
 calc_lsc_status(path_scenario, path_reference)
}

}
