% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_status.R
\name{calc_status}
\alias{calc_status}
\title{Calculate the planetary boundary status}
\usage{
calc_status(
  boundary,
  config_scenario,
  config_reference,
  spatial_scale,
  time_span_scenario = as.character(1982:2011),
  time_span_reference = time_span_scenario,
  time_aggregation_args = list(),
  approach = list(),
  thresholds = list(),
  in_parallel = TRUE,
  ...
)
}
\arguments{
\item{boundary}{character vector, boundary for which status is calculated.
Available terrestrial boundaries are c("bluewater", "greenwater", "lsc",
"nitrogen")}

\item{config_scenario}{character string. File path to the LPjmL configuration
file (json) of the scenario run. The configuration file contains the
information about the LPJmL run, e.g. the output directory}

\item{config_reference}{character string. See config_scenario. For the
reference run}

\item{spatial_scale}{character string indicating spatial resolution
options: "global", "subglobal", "grid";}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{time_aggregation_args}{list of arguments to be passed to
\code{\link{aggregate_time}} (see for more info). To be used for # nolint
time series analysis}

\item{approach}{list of methods to be used for each boundary. If \code{NULL} the
default approach is used}

\item{thresholds}{list of thresholds to be used for each boundary. If \code{NULL}
the default thresholds are used}

\item{in_parallel}{logical, if \code{TRUE} the function uses parallelization
(default) based on the \code{future} package (asynchronous execution). If \code{FALSE}
no parallelization is used}

\item{...}{further arguments to be passed to each calc_* function}
}
\value{
list with data array for each \code{boundary}
}
\description{
Calculate the PB status for a defined planetary boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{

\dontrun{
 boundary_status <- calc_status(
   boundary = "nitrogen"
   path_scenario = "./my_scenario/output")
}

}
