% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_status.R
\name{calc_status}
\alias{calc_status}
\title{Calculate the planetary boundary status}
\usage{
calc_status(
  boundary,
  spatial_scale,
  path_scenario,
  path_reference = NULL,
  time_span_scenario = as.character(1982:2011),
  time_span_reference = time_span_scenario,
  time_aggregation_args = list(),
  input_files = list(),
  diff_output_files = list(),
  method = list(),
  thresholds = list(),
  in_parallel = TRUE,
  ...
)
}
\arguments{
\item{boundary}{character vector, boundary for which status is calculated.
Available terrestrial boundaries are c("bluewater", "greenwater", "lsc",
"nitrogen")}

\item{path_scenario}{output directory (character string) of the scenario
LPJmL run where binary files (soon with metafiles) are written}

\item{path_reference}{output directory (character string) of the reference
LPJmL run where binary files (soon with metafiles) are written}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer vector, e.g. \code{1901:1930}. Can differ in offset and length from
\code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario} is used}

\item{time_aggregation_args}{list of arguments to be passed to
\link[boundaries]{aggregate_time} (see for more info). To be used for # nolint
time series analysis}

\item{input_files}{list of required file(s) using ID (e.g. \code{prec}, \code{runoff})
and an absolute file path to the corresponding input file}

\item{diff_output_files}{list of required file(s) using ID
(e.g. prec, runoff) and the alternative writing (e.g. \code{"my_runoff"})}

\item{method}{list of methods to be used for each boundary. If \code{NULL} the
default method is used}

\item{thresholds}{list of thresholds to be used for each boundary. If \code{NULL}
the default thresholds are used}

\item{...}{further arguments to be passed to each calc_* function}
}
\value{
list with data array for each \code{boundary}
}
\description{
Calculate the PB status for a defined planetary boundary based
on a scenario LPJmL run and a reference LPJmL run.
}
\examples{

\dontrun{
 boundary_status <- calc_status(
   boundary = "nitrogen"
   path_scenario = "./my_scenario/output")
}

}
