% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_status.R
\name{calc_status}
\alias{calc_status}
\title{Calculate the planetary boundary status}
\usage{
calc_status(
  boundary,
  config_scenario,
  config_reference,
  spatial_scale,
  time_span_scenario = 1982:2011,
  time_span_reference = time_span_scenario,
  time_series_avg = NULL,
  approach = list(),
  thresholds = list(),
  in_parallel = TRUE,
  ...
)
}
\arguments{
\item{boundary}{character vector, boundary for which status is calculated.
Available terrestrial boundaries are \code{c("bluewater", "greenwater", "lsc", "nitrogen", "biosphere")}.}

\item{config_scenario}{character string. File path to the LPjmL configuration
file (json) of the scenario run. The configuration file contains the
information about the LPJmL run, e.g. the output directory}

\item{config_reference}{character string. See config_scenario. For the
reference run}

\item{spatial_scale}{character string indicating spatial resolution
options: "global", "subglobal", "grid";}

\item{time_span_scenario}{time span to be used for the scenario run, defined
as an integer (or character) vector, e.g. \code{1982:2011} (default)}

\item{time_span_reference}{time span to be used for the scenario run, defined
as an integer (or character) vector, e.g. \code{1901:1930}. Can differ in offset
and length from \code{time_span_scenario}! If \code{NULL} value of \code{time_span_scenario}
is used}

\item{time_series_avg}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for one status
calculation, for \code{1} the whole time span is used to calculate a status time
series.}

\item{approach}{list of methods to be used for each boundary. If \code{NULL} the
default approach is used}

\item{thresholds}{list of thresholds to be used for each boundary. If \code{NULL}
the default thresholds are used}

\item{in_parallel}{logical, if \code{TRUE} the function uses parallelization
(default) based on the \code{future} package (asynchronous execution). If \code{FALSE}
no parallelization is used}

\item{...}{further arguments to be passed to each calc_* function}
}
\value{
list with objects of class \code{control_variable}. To directly get the
\code{boundary_status} use \code{\link[=as_risk_level]{as_risk_level()}}.
}
\description{
Calculate the PB status for a defined planetary boundary based
on a scenario LPJmL run and a reference LPJmL run. For boundary function
specific arguments to be passed (via \code{...}) see the respective function
documentation of \code{\link[=biosphere_status]{biosphere_status()}}, \code{\link[=nitrogen_status]{nitrogen_status()}},
\code{\link[=greenwater_status]{greenwater_status()}}, \code{\link[=bluewater_status]{bluewater_status()}} or \code{\link[=lsc_status]{lsc_status()}}
}
\examples{

\dontrun{
boundary_status <- calc_status(
  boundary = c("biosphere","nitrogen", "greenwater", "bluewater", "lsc")
  config_scenario = "path/to/config_scenario.json",
  config_reference = "path/to/config_reference.json",
  spatial_scale = "global",
  time_span_scenario = 1901:2019,
  time_span_reference = 1901:1930
)
}

}
