% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_biomes.R
\name{classify_biomes}
\alias{classify_biomes}
\title{Classify biomes}
\usage{
classify_biomes(
  config_reference = NULL,
  files_reference = NULL,
  time_span_reference,
  savanna_proxy = list(vegc = 7500),
  montane_arctic_proxy = list(elevation = 1000),
  tree_cover_thresholds = list(),
  approach = "default",
  nyear_window = NULL,
  config_args = list(),
  input_files = list(),
  diff_output_files = list()
)
}
\arguments{
\item{config_reference}{character string. File path to the LPjmL
configuration file (json) of the reference run. The configuration file
contains the information about the LPJmL run, e.g. the output
directory}

\item{files_reference}{list with variable names and corresponding file paths
(character string) of the reference LPJmL run. All needed files are
provided as key value pairs, e.g.:
list(leaching = "/temp/leaching.bin.json"). If path_reference is
supplied with all needed files, files reference can be set to NULL.}

\item{time_span_reference}{time span to be used for the classification of
biomes, defined as character string, e.g. \code{as.character(1901:1930)}.}

\item{savanna_proxy}{\code{list} with either pft_lai or vegc as
key and value in m2/m2 for pft_lai (default: 6) and gC/m2 for
vegc (current default: 7500); set to \code{NULL} if no proxy should be
used.}

\item{montane_arctic_proxy}{\code{list} with either "elevation" or "latitude" as
name/key and value in m for elevation (default: 1000) and degree for
latitude (default: 55); set to \code{NULL} if no proxy is used.}

\item{tree_cover_thresholds}{list with minimum tree cover thresholds for
definition of forest, woodland, savanna and grassland. Only changes to
the default have to be included in the list, for the rest the default
is used.
Default values, based on the IGBP land cover classification system:
"boreal forest" = 0.6
"temperate forest" = 0.6
"temperate woodland" = 0.3
"temperate savanna" = 0.1
"tropical forest" = 0.6
"tropical woodland" = 0.3
"tropical savanna" = 0.1
In the boreal zone, there is no woodland, everything below the
boreal forest threshold will be classified as boreal tundra.}

\item{approach}{character string indicating which biome classification
approach to use. Currently only one is defined ("default").}

\item{nyear_window}{integer. Number of years to be used for the moving
average calculation. If \code{NULL}, all years are averaged for one status
calculation, for \code{1} the whole time span is used to calculate a status
time series.}

\item{config_args}{list of arguments to be passed on from the model
configuration.}

\item{input_files}{list of required file(s) using ID (e.g. \code{temp},
\code{elevation}) and an absolute file path to the corresponding input file
if files_reference are not provided and if temp and elevation are
needed and not among the output files under path_reference}

\item{diff_output_files}{list of required file(s) using ID
(e.g. prec, runoff) and the alternative writing (e.g. \code{"my_runoff"}).
Only required if files_reference are not provided}
}
\value{
list object containing biome_id (main biome per grid
\code{cell[dim=c(ncells)]}), and list of respective
\code{biome_names[dim=c(nbiomes)]}
}
\description{
Classify biomes based on foliage protected cover (FPC) and temperature
LPJmL output plus either vegetation carbon or pft_lai depending on
the savanna_proxy option and elevation if montane_arctic_proxy requires this
}
\examples{
\dontrun{
classify_biomes(
  path_reference = "./outputs/"
  time_span_reference = c(1982:2011))
}

}
