% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomes.R
\name{plot_biomes}
\alias{plot_biomes}
\title{Plot global distribution of lpjml simulated biomes}
\usage{
plot_biomes(x, filename = NULL, projection = "+proj=robin", grid_path = NULL)
}
\arguments{
\item{x}{output (list) from classify_biomes()}

\item{filename}{directory for saving the plot (character string)}

\item{projection}{character string defining the projection, default set to
"+proj=robin"}

\item{grid_path}{character string providing the path to a grid file}
}
\description{
Plots a map with the biome distribution as derived from a lpjml run based
on the "classify_biomes" function
}
\examples{
\dontrun{

biomes <- classify_biomes(
 config_reference = path_reference,
 time_span_reference = as.character(2008:2017),
 savanna_proxy = list(vegc = 7500)
)

plot_biomes(
  x = biomes,
  filename = "/p/projects/open/Johanna/R/biomes.pfd"
  grid_path = ".grid.bin.json"
)
}

}
