% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status.R
\name{plot_status}
\alias{plot_status}
\title{Plot the status of planetary boundaries}
\usage{
plot_status(x, filename = NULL, add_legend = TRUE, stylized = FALSE, ...)
}
\arguments{
\item{x}{output object from calc_* with the status of the
control variable for one point in time, incl. pb thresholds as attribute}

\item{filename}{character string providing file name (including directory
and file extension). Defaults to NULL (plotting to screen)}

\item{add_legend}{logical, specify whether a legend should be plotted}

\item{stylized}{Logical. If \code{spatial_scale == "global"}, the function will
plot the status of the planetary boundaries using a stylized plot.}

\item{...}{additional arguments passed to the plotting functions, see also
\code{\link{plot_status_global}}, \code{\link{plot_status_maps}} and \code{\link{plot_status_stylized}}}
}
\description{
Plot the status of planetary boundaries. The function takes the output from
calc_status and plots the status of the planetary boundaries depending on
the spatial scale applying different forms of plotting.
}
\examples{
\dontrun{
pb_status <- calc_status(
  boundary = c("lsc", "biosphere", "bluewater", "greenwater", "nitrogen"),
  config_scenario = "./config_lu_1500_2016.json",
  config_reference = "./config_pnv_1500_2016.json",
  time_span_scenario = as.character(1986:2016),
  time_span_reference = as.character(1986:2016),
  spatial_scale = "global",
  approach = list(
    bluewater = "porkka2024",
    nitrogen = "schulte_uebbing2022"
  ),
  savanna_proxy = list(vegc = 7500),
  nyear_window = 1,
  path_baseline = "./pnv_1500_2016/",
)

plot_status(
  x = pb_status,
  filename = "status.png",
  add_legend = TRUE,
  stylized = TRUE
)
}

}
