% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_global.R
\name{plot_status_global}
\alias{plot_status_global}
\title{Plot the global status of planetary boundaries}
\usage{
plot_status_global(
  x,
  filename = NULL,
  all_in_one = FALSE,
  ncol = 2,
  normalize = "increasing risk"
)
}
\arguments{
\item{x}{list with global output from calc_status}

\item{filename}{character string providing file name (including directory
and file extension). Defaults to NULL (plotting to screen and
returning plot object for further customization)}

\item{all_in_one}{boolean, if TRUE, all PB stati will be normalized and
plotted in one panel}

\item{ncol}{number of plot columns (only relevant if more than one pb is
plotted and all_in_one = FALSE)}

\item{normalize}{see \code{\link[=as_risk_level]{as_risk_level()}} for details. Default set to
"increasing risk". Only relevant if all_in_one = TRUE}
}
\description{
Plot line plots with the PB status over time for
a scenario LPJmL run and derived planetary boundary statuses. Legend can be
plotted seperately based on the plot_legend() function
}
\examples{
\dontrun{
plot_status_global(
  filename = "./my_boundary_status.png",
  x = status_output,
  all_in_one = FALSE,
  ncol = 2
)
}

}
