% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_global.R
\name{plot_status_global}
\alias{plot_status_global}
\title{Plot the global status of planetary boundaries}
\usage{
plot_status_global(
  x,
  filename = NULL,
  add_legend = TRUE,
  colors = c(bluewater = "#377eb8", greenwater = "#4daf4a", lsc = "#ff7f00", biosphere =
    "#e41a1c", nitrogen = "#984ea3"),
  all_in_one = FALSE,
  ncol = 2,
  normalize = "safe"
)
}
\arguments{
\item{x}{list with global output from calc_status}

\item{filename}{character string providing file name (including directory
and file extension). Defaults to NULL (plotting to screen)}

\item{add_legend}{logical, specify whether a legend should be plotted}

\item{colors}{named vector for definition of colors for plotting of
the safe, increasingnrisk and high risk zone}

\item{all_in_one}{boolean, if TRUE, all PB stati will be normalized to 0
(holocene value) to 1 (pb) and plotted in one panel}

\item{ncol}{number of plot columns (only relevant if more than one pb is
plotted and all_in_one = FALSE)}

\item{normalize}{character string to define normalization, either "safe"
(normalized from holocene to pb = the safe zone) or
"increasing risk" (normalized from pb to high risk level =
increasing risk zone if the pb status is > pb, otherwise normalized
from holocene to pb). Only used if all_on_one set to TRUE}
}
\description{
Plot line plots with the PB status over time  for
a scenario LPJmL run and derived planetary boundary statuses
}
\examples{
\dontrun{
plot_status(
  filename = "./my_boundary_status.png",
  x = list(
    "land system change" = lsc_status,
    "nitrogen" = nitrogen_status,
    "bluewater" = water_status
  ),
  legend = FALSE,
  bg_col = NA
)
}

}
